DjangoBoard/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.7
django-bootstrap5==23.3
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
mysqlclient==2.2.0
packaging==23.2
platformdirs==4.0.0
pytz==2023.3.post1
qbittorrent-api==2023.11.57
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.25.0
